"The scientific man does not aim at an immediate result. He does not expect that his advanced ideas will be readily taken up. His work is like that of the planter — for the future. His duty is to lay the foundation for those who are to come, and point the way. He lives and labors and hopes." - Nikola Tesla
This is a new branch for the new posting system, things will be broken. From it will come new features and capabilities, greater performance and, most importantly, a level of simplicity not offered by the previous system. Functionally similar but widely different in terms of implementation, ease of use and extendability...


git-svn-id: file:///svn/phpbb/branches/posting@8706 89ea8834-ac86-4346-8a33-228a782c2dd0